reorganize data loader folders
