+++ asr/data/asr_dataset/asr_dataset/constants.py
@@ -9,5 +9,6 @@ DATASET_DIRS = {
         'radio_transcripts':'/net/projects/uri/transcripts',
         'radio_mp3s':'/net/projects/uri/data',
         'librispeech':'/net/projects/uri/ra/shiyanglai/experiment1/audio data/LibriSpeech/train-clean-100',
+        'atc0': '/net/projects/uri/data/corpora/atc0_comp',
     }
 }
