+++ asr/data/asr_dataset/asr_dataset/radio.py
@@ -9,8 +9,8 @@ import re
 import datetime
 import pandas as pd
 import logging
-from asr_dataset.dataset import AudioClipDataset
-from asr_dataset.datasets.constants import DATASET_DIRS
+from asr_dataset.base_dataset import AudioClipDataset
+from asr_dataset.constants import DATASET_DIRS
 
 BAD_WORDS = ["\[UNCERTAIN\]", "<X>", "INAUDIBLE"] # used as regex, thus [] escaped
 SAMPLE_RATE = 16000  # Hz
